+++ pdiiif-web/tailwind.config.js
@@ -1,5 +1,16 @@
 module.exports = {
   content: ['./index.html', './src/**/*.svelte'],
-  theme: { extend: { colors: { brand: '#ee2635' }}},
+  theme: { 
+    extend: { 
+      colors: { 
+        brand: '#001158',
+        'bg-1': '#001158',
+        'bg-2': '#001158',
+        'card-bg': '#001158',
+        'accent': '#001158',
+        'muted': 'rgba(255,255,255,0.72)',
+      }
+    }
+  },
   plugins: [],
 };